requirements.txt
fire==0.5.0
python-magic==0.4.27
termcolor==1.1.0
